requirements.txt
python-telegram-bot
yt-dlp
python-dotenv


